REFACTOR: move generic functions to utils.ts
